Tag EmbeddingVectorConverter 수정 (#85)

+++ src/main/kotlin/com/wafflestudio/toyproject/memoWithTags/tag/dto/TagRequest.kt
@@ -8,7 +8,7 @@ sealed class TagRequest {
         val id: UUID,
         val name: String,
         val colorHex: String,
-        val embeddingVector: List<Float>,
+        val embeddingVector: List<Double>,
         val createdAt: Instant,
         val updatedAt: Instant?
     ) : TagRequest()
@@ -17,7 +17,7 @@ sealed class TagRequest {
         val id: UUID,
         val name: String,
         val colorHex: String,
-        val embeddingVector: List<Float>,
+        val embeddingVector: List<Double>,
         val createdAt: Instant,
         val updatedAt: Instant?
     ) : TagResponse()
